move MeshGL to its own header
